Rename namespaces to match directories.
