+++ PrimeCSharp/solution_1/V1Sieves/PrimeSieve.cs
@@ -2,7 +2,7 @@
 using System.Collections;
 using System.Collections.Generic;
 
-namespace PrimeCSharp.Sieves
+namespace PrimeCSharp.V1Sieves
 {
     class PrimeSieve : ISieve
     {
