+++ PrimeCSharp/solution_1/V1Sieves/PrimeSieveArrayPool.cs
@@ -7,7 +7,7 @@ using System.Runtime.CompilerServices;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace PrimeCSharp.Sieves
+namespace PrimeCSharp.V1Sieves
 {
     class PrimeSieveArrayPool : ISieve
     {
