+++ PrimeCSharp/solution_1/V1Sieves/PrimeSieveArrayPool8of30.cs
@@ -3,7 +3,7 @@ using System.Buffers;
 using System.Collections.Generic;
 using System.Runtime.CompilerServices;
 
-namespace PrimeCSharp.Sieves
+namespace PrimeCSharp.V1Sieves
 {
     class PrimeSieveArrayPool8of30 : ISieve
     {
