+++ PrimeCSharp/solution_1/V1Sieves/PrimeSieveBool.cs
@@ -1,7 +1,7 @@
 ﻿using System;
 using System.Collections.Generic;
 
-namespace PrimeCSharp.Sieves
+namespace PrimeCSharp.V1Sieves
 {
     class PrimeSieveBool : ISieve
     {
