+++ PrimeCSharp/solution_1/V1Sieves/PrimeSieveInvBoolB.cs
@@ -1,7 +1,7 @@
 ﻿using System;
 using System.Collections.Generic;
 
-namespace PrimeCSharp.Sieves
+namespace PrimeCSharp.V1Sieves
 {
     class PrimeSieveInvBoolB : ISieve
     {
