+++ PrimeCSharp/solution_1/V1Sieves/PrimeSieveRawParallel.cs
@@ -4,7 +4,7 @@ using System.Runtime.CompilerServices;
 using System.Threading;
 using System.Threading.Tasks;
 
-namespace PrimeCSharp.Sieves
+namespace PrimeCSharp.V1Sieves
 {
     class PrimeSieveRawParallel : ISieve
     {
