+++ PrimeCSharp/solution_1/V2Sieves/Bit2Sieve.cs
@@ -2,7 +2,7 @@
 using System.Collections;
 using System.Collections.Generic;
 
-namespace PrimeCSharp.SieveRunners
+namespace PrimeCSharp.V2Sieves
 {
     public class Bit2Sieve : ISieveRunner
     {
