+++ PrimeCSharp/solution_1/V2Sieves/IBool30Sieve.cs
@@ -2,7 +2,7 @@
 using System.Collections;
 using System.Collections.Generic;
 
-namespace PrimeCSharp.SieveRunners
+namespace PrimeCSharp.V2Sieves
 {
     public class IBool30Sieve : ISieveRunner
     {
