+++ PrimeCSharp/solution_1/V2Sieves/PoolB2Sieve.cs
@@ -4,7 +4,7 @@ using System.Collections;
 using System.Collections.Generic;
 using System.Runtime.CompilerServices;
 
-namespace PrimeCSharp.SieveRunners
+namespace PrimeCSharp.V2Sieves
 {
     public class PoolB2Sieve : ISieveRunner
     {
