+++ PrimeCSharp/solution_1/V2Sieves/PoolQ6Sieve.cs
@@ -4,7 +4,7 @@ using System.Collections;
 using System.Collections.Generic;
 using System.Runtime.CompilerServices;
 
-namespace PrimeCSharp.SieveRunners
+namespace PrimeCSharp.V2Sieves
 {
     public class PoolQ6Sieve : ISieveRunner
     {
