+++ PrimeCSharp/solution_1/V2Sieves/RawD2Sieve.cs
@@ -5,7 +5,7 @@ using System.Collections.Generic;
 using System.Runtime.CompilerServices;
 using System.Runtime.InteropServices;
 
-namespace PrimeCSharp.SieveRunners
+namespace PrimeCSharp.V2Sieves
 {
     public class RawD2Sieve : ISieveRunner
     {
